chainlit-groq-demo/pyproject.toml
[project]
name = "farm-control-demo"
version = "1.0.0"
description = "Farm Control Demo with Chainlit and GROQ API"
requires-python = ">=3.10"

dependencies = [
    "chainlit>=1.0.0",
    "groq>=0.9.0",
    "mcp==1.9.4",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.1",
    "typing-extensions>=4.9.0",
    "sqlalchemy>=2.0.25",
    "requests>=2.31.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "livekit-agents>=0.4.1",
    "openai-whisper>=20250625",
    "gtts>=2.5.4",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "pyaudio>=0.2.14",
    "pyttsx3>=2.98",
    "langchain-core>=0.3.66",
    "langchain>=0.3.26",
    "langgraph>=0.5.0",
    "langchain-openai>=0.3.27",
    "langchain-community>=0.3.26",
    "langchain-text-splitters>=0.3.8",
    "langchain-mcp>=0.2.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "farm_control_server"]
